+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * **login, signup and otp component:** modified login, added signup and otp (wip) ([9fcb4b3](https://gitlab.com/paras205/boilerplate-react/commit/9fcb4b32b0c646f71ab088749a63b887740c16f3))
 * **registration:** added stepper form components for registration ([f00b3fd](https://gitlab.com/paras205/boilerplate-react/commit/f00b3fd93b4a8e41c082fef6d63ddca36699bd38))
 * **sign up:** changes in styles,removed component,otp input,api ([079389d](https://gitlab.com/paras205/boilerplate-react/commit/079389d3d93e597a0be7a35fb0a013cdb751aec7))
+* **stepper form field name:** changes in form field name ([fa90159](https://gitlab.com/paras205/boilerplate-react/commit/fa90159f2e8de00e1ad1fcd7d797694d3126d34c))
+* **stepper forms:** created stepper forms and adjusted styles ([db8e878](https://gitlab.com/paras205/boilerplate-react/commit/db8e87873ff7993fdcf27f2bb2052e1f25b9f253))